DMS/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
mysql-connector-python==8.1.0
pydantic==2.5.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.31.0
twilio==8.13.0
httpx==0.25.2
alembic==1.12.1
python-multipart==0.0.6
typing-extensions==4.8.0